+++ lib/src/models/payment_status/payment_status.dart
@@ -1,0 +1,11 @@
+///
+enum PaymentStatus {
+  ///
+  cancel,
+
+  ///
+  failure,
+
+  ///
+  success,
+}