+++ lib/src/paymaya.dart
@@ -1,12 +1,19 @@
 import 'dart:convert';
+
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 import 'package:http/http.dart';
-import 'package:paymaya_flutter/src/models/response/response.dart';
 
 import 'models/models.dart';
+import 'models/payment_status/payment_status.dart';
+import 'widget/checkout_page.dart';
+import 'widget/single_payment_page.dart';
+
+final _argumentError =
+    ArgumentError('Initialize PayMaya SDK using PayMayaSDK.init');
 
 /// {@template paymayasdk}
-///
+///z
 /// This SDK is a mirror of [PayMayaSDK-JS](https://github.com/PayMaya/PayMaya-JS-SDK-v2).
 ///
 /// The author chose this because other SDK uses WebView and that is not ideal
@@ -26,9 +33,38 @@ abstract class PayMayaSDK {
   ///
   /// Throws [ArgumentError] when invoking other methods
   /// without running `init()`.
-  factory PayMayaSDK.init(String publicKey, {bool isSandbox = true}) {
-    return _PayMayaSDK(publicKey, isSandbox);
+  /// If you want to customize your authentication process.
+  /// you can pass `http` as your client and have different implementation.
+  /// ```dart
+  /// class _PayMayaHttp extends BaseClient {
+  ///   _PayMayaHttp(this.publicKey);
+  ///   final String publicKey;
+  ///   final Client _client = Client();
+  ///   @override
+  ///   Future<StreamedResponse> send(BaseRequest request) {
+  ///     final bytes = utf8.encode(publicKey);
+  ///     final base64Str = base64.encode(bytes);
+  ///    final headers = {
+  ///       'Content-Type': 'application/json',
+  ///       'Authorization': 'Basic $base64Str',
+  ///     };
+  ///     request.headers.addAll(headers);
+  ///     return _client.send(request);
+  ///   }
+  /// }
+  /// ```
+  ///
+
+  factory PayMayaSDK.init(
+    String publicKey, {
+    bool isSandbox = true,
+    BaseClient? http,
+  }) {
+    return _PayMayaSDK(publicKey, isSandbox, http);
   }
+
+  ///
+  bool get hasPublicKey => false;
 
   /// Use this to request request link from PayMaya API.
   /// Be sure to read the documentation
@@ -51,27 +87,29 @@ abstract class PayMayaSDK {
     throw ArgumentError('Initialize PayMaya SDK using PayMayaSDK.init');
   }
 
+  /// Consider using [navigateToCheckout] for easier integration.
+  ///
   /// This method redirects the user to PayMaya Checkout, where the user can
   /// finalize his/her payment.
   ///
   /// This method returns a url which you will have to use whatever technique
   /// you can use(url_launcher, webview, or iframe).
   Future<PaymayaResponse> createCheckOut(PaymayaCheckout checkout) {
-    throw ArgumentError('Initialize PayMaya SDK using PayMayaSDK.init');
+    throw _argumentError;
   }
 
   /// creates a wallet link
   /// that allows charging to a PayMaya account.
   Future<PaymayaResponse> createWalletLink(
       CreateWalletObject createWalletObject) {
-    throw ArgumentError('Initialize PayMaya SDK using PayMayaSDK.init');
+    throw _argumentError;
   }
 
   /// creates a single payment redirection,
   /// allowing the user to finalize the transaction.
   Future<PaymayaResponse> createSinglePayment(
       PaymayaSinglePayment createSinglePayment) {
-    throw ArgumentError('Initialize PayMaya SDK using PayMayaSDK.init');
+    throw _argumentError;
   }
 
   /// ### Not suitable for production. IF you have suggestions, give us a PR.
@@ -79,6 +117,21 @@ abstract class PayMayaSDK {
   /// and fills up the credit card form.
   @visibleForTesting
   Future<String> createdCreditCardForm() {
+    throw _argumentError;
+  }
+
+  /// Processes your transaction automatically
+  /// to use this in all platforms, consider doing this.
+  ///
+  Future<PaymentStatus> navigateToCheckout(
+      BuildContext context, PaymayaCheckout checkout) {
+    throw _argumentError;
+  }
+
+  /// Handles your payment automatically.
+  ///
+  Future<PaymentStatus> navigateToPayment(
+      BuildContext context, PaymayaSinglePayment checkout) {
     throw ArgumentError('Initialize PayMaya SDK using PayMayaSDK.init');
   }
 }
@@ -87,7 +140,9 @@ class _PayMayaSDK implements PayMayaSDK {
   _PayMayaSDK(
     this.publicKey,
     this.isSandbox,
+    BaseClient? http,
   ) {
+    _http = http ?? _PayMayaHttp(publicKey);
     if (isSandbox) {
       apiUrl = 'https://pg-sandbox.paymaya.com';
       formUrl = 'https://paymayajs-staging.s3.amazonaws.com/dist/index.html';
@@ -104,11 +159,26 @@ class _PayMayaSDK implements PayMayaSDK {
   late final bool isSandbox;
   late String formUrl;
   late String eventOrigin;
-
+  late BaseClient _http;
   void _checkIfInitialize() {
     if ((publicKey.isEmpty)) {
       throw ArgumentError('Public Key must not be empty');
     }
+  }
+
+  @override
+  Future<Map<String, dynamic>> genericRequestFn(
+      {required Map<String, dynamic> requestBody, required String url}) async {
+    final response = await _http.post(
+      Uri.parse('$apiUrl$url'),
+      body: jsonEncode(requestBody),
+    );
+    final Map<String, dynamic> json = jsonDecode(response.body);
+    print(json);
+    if (response.statusCode != 200 || json['redirectUrl'] == null) {
+      throw Exception(json.toString());
+    }
+    return json;
   }
 
   @override
@@ -174,19 +244,42 @@ class _PayMayaSDK implements PayMayaSDK {
   }
 
   @override
-  Future<Map<String, dynamic>> genericRequestFn(
-      {required Map<String, dynamic> requestBody, required String url}) async {
-    final _http = _PayMayaHttp(publicKey);
-    final response = await _http.post(
-      Uri.parse('$apiUrl$url'),
-      body: jsonEncode(requestBody),
+  bool get hasPublicKey => publicKey.isNotEmpty;
+
+  @override
+  Future<PaymentStatus> navigateToCheckout(
+      BuildContext context, PaymayaCheckout checkout) async {
+    final response = await createCheckOut(checkout);
+    final PaymentStatus status = await Navigator.push(
+      context,
+      CupertinoPageRoute(
+        builder: (context) {
+          return PaymayaCheckoutPage(
+            url: response.redirectUrl,
+            redirectUrls: checkout.redirectUrl,
+          );
+        },
+      ),
     );
-    final Map<String, dynamic> json = jsonDecode(response.body);
-    print(json);
-    if (response.statusCode != 200 || json['redirectUrl'] == null) {
-      throw Exception(json.toString());
-    }
-    return json;
+    return status;
+  }
+
+  @override
+  Future<PaymentStatus> navigateToPayment(
+      BuildContext context, PaymayaSinglePayment singlePayment) async {
+    final response = await createSinglePayment(singlePayment);
+    final PaymentStatus status = await Navigator.push(
+      context,
+      CupertinoPageRoute(
+        builder: (context) {
+          return PaymayaSinglePaymentPage(
+            url: response.redirectUrl,
+            redirectUrls: singlePayment.redirectUrl,
+          );
+        },
+      ),
+    );
+    return status;
   }
 }