+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.2.0]
+### Features
+- Added Checkout Widget. Use `PayMayaSDK.navigateToCheckout(PaymayaCheckout checkout)`.
+- Added Single Payment Widget. Use `PayMayaSDK.navigateToPayment(PaymayaSinglePayment singlePayment)`
+- Added Link widget that uses `url_launcher` Platform specific webview including Web, Android and iOS. Use `PaymayaLink` Widget. See [PaymayaLink](./lib/src/widgets/link.dart)
+### Breaking Changes
+- Change `createCheckout` and `createSinglePayment` return type to `PaymayaResponse`. See [PaymayaResponse](./lib/src/models/response.dart)
+- Added `http` on `PayMayaSDK.init` method. This lets you customize your http response.
+
 ## [0.1.0]
 ### Features